scripts/voice-agent-requirements.txt
twilio==9.0.0
openai==1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
